Refactoring Index file to User and Product files for seperate functions
